feat(api): also save nonAsso siren and names
