+++ packages/api/src/modules/providers/sirene/stockUniteLegale/sireneStockUniteLegale.file.service.test.ts
@@ -1,4 +1,4 @@
-import sireneStockUniteLegaleService from "./sireneStockUniteLegale.service";
+import sireneStockUniteLegaleFileService from "./sireneStockUniteLegale.file.service";
 import sireneStockUniteLegaleApiPort from "../../../../dataProviders/api/sirene/sireneStockUniteLegale.port";
 import sireneStockUniteLegaleDbPort from "../../../../dataProviders/db/sirene/stockUniteLegale/sireneStockUniteLegale.port";
 import { Readable } from "stream";
@@ -36,22 +36,22 @@ describe("SireneStockUniteLegaleService", () => {
     describe("getOrCreateDirectory", () => {
         it("should check if the directory exists", () => {
             // @ts-expect-error : we are testing a private method
-            sireneStockUniteLegaleService.getOrCreateDirectory();
+            sireneStockUniteLegaleFileService.getOrCreateDirectory();
             // @ts-expect-error : private variable
-            expect(fs.existsSync).toHaveBeenCalledWith(sireneStockUniteLegaleService.directory_path);
+            expect(fs.existsSync).toHaveBeenCalledWith(sireneStockUniteLegaleFileService.directory_path);
         });
 
         it("should create a directory if it does not exist", () => {
             jest.mocked(fs.existsSync).mockReturnValueOnce(false);
             // @ts-expect-error : we are testing a private method
-            sireneStockUniteLegaleService.getOrCreateDirectory();
+            sireneStockUniteLegaleFileService.getOrCreateDirectory();
             expect(fs.mkdtempSync).toHaveBeenCalledWith(expect.stringContaining("/tmpSirene"));
         });
 
         it("should not create a directory if it exists", () => {
             jest.mocked(fs.existsSync).mockReturnValueOnce(true);
             // @ts-expect-error : we are testing a private method
-            sireneStockUniteLegaleService.getOrCreateDirectory();
+            sireneStockUniteLegaleFileService.getOrCreateDirectory();
             expect(fs.mkdtempSync).not.toHaveBeenCalled();
         });
     });
@@ -62,11 +62,11 @@ describe("SireneStockUniteLegaleService", () => {
         let deleteTemporaryFolderMock: jest.SpyInstance;
         beforeAll(() => {
             getExtractAndSaveFilesMock = jest
-                .spyOn(sireneStockUniteLegaleService, "getExtractAndSaveFiles")
+                .spyOn(sireneStockUniteLegaleFileService, "getExtractAndSaveFiles")
                 .mockResolvedValue();
             parseCsvAndInsertMock = jest.spyOn(SireneStockUniteLegaleParser, "parseCsvAndInsert").mockResolvedValue();
             deleteTemporaryFolderMock = jest
-                .spyOn(sireneStockUniteLegaleService, "deleteTemporaryFolder")
+                .spyOn(sireneStockUniteLegaleFileService, "deleteTemporaryFolder")
                 .mockReturnValue();
         });
 
@@ -75,20 +75,20 @@ describe("SireneStockUniteLegaleService", () => {
         });
 
         it("should call getExtractAndSaveFiles", async () => {
-            await sireneStockUniteLegaleService.getAndParse();
+            await sireneStockUniteLegaleFileService.getAndParse();
             expect(getExtractAndSaveFilesMock).toHaveBeenCalledTimes(1);
         });
 
         it("should call parseCsvAndInsert", async () => {
-            await sireneStockUniteLegaleService.getAndParse();
+            await sireneStockUniteLegaleFileService.getAndParse();
             expect(parseCsvAndInsertMock).toHaveBeenCalledWith(
                 // @ts-expect-error : private variable
-                sireneStockUniteLegaleService.directory_path + "/StockUniteLegale_utf8.csv",
+                sireneStockUniteLegaleFileService.directory_path + "/StockUniteLegale_utf8.csv",
             );
         });
 
         it("should call deleteTemporaryFolder", async () => {
-            await sireneStockUniteLegaleService.getAndParse();
+            await sireneStockUniteLegaleFileService.getAndParse();
             expect(deleteTemporaryFolderMock).toHaveBeenCalledTimes(1);
         });
     });
@@ -139,17 +139,17 @@ describe("SireneStockUniteLegaleService", () => {
         });
 
         it("should call createWriteStream", async () => {
-            await sireneStockUniteLegaleService.getAndSaveZip();
+            await sireneStockUniteLegaleFileService.getAndSaveZip();
             expect(fs.createWriteStream).toHaveBeenCalledWith(expect.stringContaining("SireneStockUniteLegale.zip"));
         });
 
         it("should call getZip", async () => {
-            await sireneStockUniteLegaleService.getAndSaveZip();
+            await sireneStockUniteLegaleFileService.getAndSaveZip();
             expect(sireneStockUniteLegaleApiPort.getZip).toHaveBeenCalledTimes(1);
         });
 
         it("should download and write the data to the file without errors", async () => {
-            const acutal = await sireneStockUniteLegaleService.getAndSaveZip();
+            const acutal = await sireneStockUniteLegaleFileService.getAndSaveZip();
             expect(acutal).toBe("finish");
         });
 
@@ -163,7 +163,7 @@ describe("SireneStockUniteLegaleService", () => {
                 status: 300,
                 statusText: "Not ok",
             });
-            await expect(sireneStockUniteLegaleService.getAndSaveZip()).rejects.toThrow(
+            await expect(sireneStockUniteLegaleFileService.getAndSaveZip()).rejects.toThrow(
                 "simulated error during reading",
             );
         });
@@ -189,7 +189,7 @@ describe("SireneStockUniteLegaleService", () => {
 
             (fs.createWriteStream as jest.Mock).mockReturnValue(mockFileStream);
 
-            await expect(sireneStockUniteLegaleService.getAndSaveZip()).rejects.toThrow(
+            await expect(sireneStockUniteLegaleFileService.getAndSaveZip()).rejects.toThrow(
                 "simulated error during writing",
             );
         });
@@ -197,17 +197,17 @@ describe("SireneStockUniteLegaleService", () => {
 
     describe("decompressFolder", () => {
         it("should call StreamZip", async () => {
-            await sireneStockUniteLegaleService.decompressFolder(ZIP_PATH, DIRECTORY_PATH);
+            await sireneStockUniteLegaleFileService.decompressFolder(ZIP_PATH, DIRECTORY_PATH);
             expect(StreamZip.async).toHaveBeenCalledWith({ file: ZIP_PATH });
         });
 
         it("should call extract", async () => {
-            await sireneStockUniteLegaleService.decompressFolder(ZIP_PATH, DIRECTORY_PATH);
+            await sireneStockUniteLegaleFileService.decompressFolder(ZIP_PATH, DIRECTORY_PATH);
             expect(StreamZip.async).toHaveBeenCalledWith({ file: ZIP_PATH });
         });
 
         it("should call close", async () => {
-            await sireneStockUniteLegaleService.decompressFolder(ZIP_PATH, DIRECTORY_PATH);
+            await sireneStockUniteLegaleFileService.decompressFolder(ZIP_PATH, DIRECTORY_PATH);
             expect(StreamZip.async).toHaveBeenCalledWith({ file: ZIP_PATH });
         });
     });
@@ -215,8 +215,8 @@ describe("SireneStockUniteLegaleService", () => {
     describe("deleteTemporaryFolder", () => {
         it("should call fs.rmdirSync", () => {
             // @ts-expect-error : private variable
-            sireneStockUniteLegaleService.directory_path = DIRECTORY_PATH;
-            sireneStockUniteLegaleService.deleteTemporaryFolder();
+            sireneStockUniteLegaleFileService.directory_path = DIRECTORY_PATH;
+            sireneStockUniteLegaleFileService.deleteTemporaryFolder();
 
             expect(fs.rmSync).toHaveBeenCalledWith(DIRECTORY_PATH, { recursive: true });
         });
@@ -233,7 +233,7 @@ describe("SireneStockUniteLegaleService", () => {
 
         it("should call insertOne", async () => {
             const dbo = { siren: "123456789" } as unknown as SireneUniteLegaleDbo;
-            await sireneStockUniteLegaleService.insertOne(dbo);
+            await sireneStockUniteLegaleFileService.insertOne(dbo);
             expect(insertOneMock).toHaveBeenCalledWith(dbo);
         });
     });
@@ -249,7 +249,7 @@ describe("SireneStockUniteLegaleService", () => {
 
         it("should call insertMany", async () => {
             const dbos = [{ siren: "123456789" }] as unknown as SireneUniteLegaleDbo[];
-            await sireneStockUniteLegaleService.insertMany(dbos);
+            await sireneStockUniteLegaleFileService.insertMany(dbos);
             expect(insertManyMock).toHaveBeenCalledWith(dbos);
         });
     });
